# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Workforce Analytics & IT Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leave-balances:
    get:
      operationId: listLeaveBalances
      tags: [leave]
      summary: List leave balances
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Leave balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveBalanceItem"
    post:
      operationId: upsertLeaveBalance
      tags: [leave]
      summary: Create or update a leave balance allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertLeaveBalanceRequest"
      responses:
        "200":
          description: Leave balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveBalanceItem"
  /leave-balances/{id}:
    delete:
      operationId: deleteLeaveBalance
      tags: [leave]
      summary: Delete a leave balance allocation
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /managers:
    get:
      operationId: listManagers
      tags: [managers]
      summary: List this tenant's managers and team members
      responses:
        "200":
          description: Managers and team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyUser"
    post:
      operationId: createManager
      tags: [managers]
      summary: Create a manager or team member in this tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManagerRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUser"
  /managers/{id}:
    patch:
      operationId: updateManager
      tags: [managers]
      summary: Update a manager or team member in this tenant
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateManagerRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUser"
    delete:
      operationId: deleteManager
      tags: [managers]
      summary: Remove a manager or team member from this tenant
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    OkResult:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    CompanyUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_user, company_admin, manager, team_member]
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, role]
...
    CreateManagerRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [manager, team_member]
      required: [username, email, password]
    UpdateManagerRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [manager, team_member]
...
    LeaveBalanceItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: [string, "null"]
        year:
          type: integer
        leaveType:
          type: string
          enum: [annual, sick, casual, unpaid]
        allocatedDays:
          type: number
        usedDays:
          type: number
        remainingDays:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - year
        - leaveType
        - allocatedDays
        - usedDays
        - remainingDays
        - createdAt
        - updatedAt
    UpsertLeaveBalanceRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        year:
          type: integer
        leaveType:
          type: string
          enum: [annual, sick, casual, unpaid]
        allocatedDays:
          type: number
      required: [userId, year, leaveType, allocatedDays]
...